Add scripts for reference machine benchmarking: (#94)
* ./scripts/reference_machine_run_all.sh
  * Sets CPU frequency and io profiles to simulate an AWS c7a.2xlarge instance with EBS io2 8000 IOPS storage.
  * Script has to be run as root/with sudo.
  * Script detects if it's running on the benchmark machine and can only be run there.
  * Machine CPU frequency settings will be restored to defaults after execution and temporary storage is cleaned up.
  * Script sets up a simulated block device stored in memory (/dev/shm) with performance throttled, this ensures consistent performance between runs.
* ./scripts/run_all_benchmarks.sh and ./scripts/bench.sh
  * Scripts were modified to allow setting --base-path argument for 'nh-node benchmark'.
  * support was added to build the docker image automatically with '--features=runtime-benchmarks'.
  * Docker Compose is used to set I/O limits, 'docker run' CLI switches to set I/O limits are not working due to a bug in docker upstream.
  * Two compose yml files were added, one for running on developers local machine, one for running on the benchmark machine with I/O throttling.

+++ scripts/run_all_benchmarks.sh
@@ -19,44 +19,60 @@
 #
 # Should be run on a reference machine to gain accurate benchmarks
 
-# shellcheck disable=SC2181
+set -eEuo pipefail
 
-ECHO_CMD=${ECHO_CMD:="false"}
+ECHO_CMD="${ECHO_CMD:-false}"
+[ "${ECHO_CMD}" = "true" ] && set -x
 
-if [ "${ECHO_CMD}" == "true" ]; 
-then
-      set -x
-fi
+# The following line ensures we know the project root
+SOURCE_ROOT="${SOURCE_ROOT:-$(git rev-parse --show-toplevel)}"
 
-USE_DOCKER_IMAGE=${USE_DOCKER_IMAGE:=""}
-# Force specific set of pallets (spaced string)
-USE_PALLETS=${USE_PALLETS:=""}
-ENABLE_OVERHEAD=${ENABLE_OVERHEAD:="true"}
-ENABLE_MACHINE=${ENABLE_MACHINE:="true"}
+# shellcheck source=/dev/null
+source "${SOURCE_ROOT}/scripts/bench_cfg.sh"
 
-# The following line ensure we know the project root
-SOURCE_ROOT=${SOURCE_ROOT:-$(git rev-parse --show-toplevel)}
+USE_DOCKER="${USE_DOCKER:-false}"
+ENABLE_PALLETS="${ENABLE_PALLETS:-false}"
+ENABLE_OVERHEAD="${ENABLE_OVERHEAD:-true}"
+ENABLE_MACHINE="${ENABLE_MACHINE:-true}"
+BENCH_BASE_PATH="${BENCH_BASE_PATH:-}"
+BASE_PATH_ARG=""
+[ -n "${BENCH_BASE_PATH:-}" ] && BASE_PATH_ARG="--base-path=${BENCH_BASE_PATH}"
 
-. "${SOURCE_ROOT}/scripts/bench_cfg.sh"
-
-BENCH_SH=${BENCH_SH:-"${SOURCE_ROOT}/scripts/bench.sh"}
+BENCH_SH="${BENCH_SH:-${SOURCE_ROOT}/scripts/bench.sh}"
 # Define the error file.
-ERR_FILE=${ERR_FILE:-"benchmarking_errors.txt"}
+ERR_FILE="${ERR_FILE:-${SOURCE_ROOT}/benchmarking_errors.txt}"
 
-if [ -z "${USE_DOCKER_IMAGE}" ]; 
-then
-echo "[+] Compiling nh-node benchmarks..."
-cargo build \
-  --locked \
-  --features=runtime-benchmarks \
-  --bin nh-node
+if [ "${USE_DOCKER}" = "false" ]; then
+  echo "[+] Compiling nh-node benchmarks..."
+  cargo build \
+    --locked \
+    --features=runtime-benchmarks \
+    --profile=production \
+    --bin nh-node
 
   # The executable to use.
   NH_NODE="${PROJECT_ROOT}/target/debug/nh-node"
   SKIP_LINES=2
 else
+  IMAGE="zkverify"
+  TAG="$(git rev-parse --short HEAD)"
+  dirty="$(git status --porcelain --untracked-files=no | grep "Cargo\|docker/\|native/\|node/\|pallets/\|primitives/\|rpc/\|verifiers/\|runtime/" | grep -v "runtime/src/weights" || true)"
+  build="false"
+  [ -n "${dirty:-}" ] && build="true" && TAG="${TAG}-dirty"
+  [ -z "$(docker image ls -q "${IMAGE}:${TAG}")" ] && build="true"
+  compose_file="${SOURCE_ROOT}/scripts/docker-compose-local.yml"
+  [ "${IS_BENCHMACHINE:-}" = "true" ] && compose_file="${SOURCE_ROOT}/scripts/docker-compose-bench.yml"
+  USER_ID="$(id -u)"
+  GROUP_ID="$(id -g)"
+  export IMAGE TAG SOURCE_ROOT USER_ID GROUP_ID ECHO_CMD
+
+  if [ "${build}" = "true" ]; then
+    echo -e "[+] Building ${IMAGE}:${TAG} runtime-benchmarks docker image...\n\n"
+    docker compose --progress=plain -f "${compose_file}" build
+    docker image prune -f
+  fi
   # The executable to use.
-  NH_NODE="docker run -ti --rm -v ${PROJECT_ROOT}:/data/benchmark --entrypoint /usr/local/bin/nh-node ${USE_DOCKER_IMAGE}"
+  NH_NODE="docker compose -f ${compose_file} run -T --rm --remove-orphans zkverify-bench /usr/local/bin/nh-node"
 
   # Now PROJECT_ROOT become the docker folder
   PROJECT_ROOT="/data/benchmark"
@@ -65,18 +81,15 @@ fi
 
 DEFAULT_DEPLOY_WEIGHT_TEMPLATE="${PROJECT_ROOT}/node/hl-deploy-weight-template.hbs"
 
-WEIGTH_TEMPLATE=${WEIGTH_TEMPLATE:-"${DEFAULT_DEPLOY_WEIGHT_TEMPLATE}"}
-WEIGHTS_FOLDER=${WEIGHTS_FOLDER:-"${PROJECT_ROOT}/runtime/src/weights"}
+WEIGTH_TEMPLATE="${WEIGTH_TEMPLATE:-${DEFAULT_DEPLOY_WEIGHT_TEMPLATE}}"
+WEIGHTS_FOLDER="${WEIGHTS_FOLDER:-${PROJECT_ROOT}/runtime/src/weights}"
 
 CODE_HEADER="${PROJECT_ROOT}/HEADER-APACHE2"
 
-
-if [ -n "${USE_PALLETS}" ]; then
-  # shellcheck disable=SC2206
-  PALLETS=(${USE_PALLETS})
-else 
+declare -a PALLETS=()
+if [ "${ENABLE_PALLETS:-}" = "true" ]; then
   # Load all pallet names in an array.
-  mapfile -t < <(${NH_NODE} benchmark pallet \
+  mapfile -t PALLETS < <(${NH_NODE} benchmark pallet \
     --list \
     --chain=dev | \
     tail -n+${SKIP_LINES} | \
@@ -84,9 +97,7 @@ else
     sort | \
     uniq \
     )
-  PALLETS=("${MAPFILE[@]}")
 fi
-
 
 EXCLUDED_PALLETS=(
         # Helper pallets
@@ -99,9 +110,9 @@ EXCLUDED_PALLETS=(
         # Not applicable now
         "pallet_session" # Crash [to investigate]
         "pallet_staking" # Not applicable if we didn't use pallet_bag_list for VoterList and TargetList
-                         # UseNominatorsAndValidatorsMap and UseValidatorsMap dosn't implement benchmark 
+                         # UseNominatorsAndValidatorsMap and UseValidatorsMap doesn't implement benchmark
                          # support
-        # SLOW 
+        # SLOW
         # "pallet_im_online" "frame_benchmarking" "frame_system" "pallet_balances"
     )
 
@@ -112,16 +123,16 @@ is_pallet_excluded() {
   local pallet=$1;
 
   for exluded in "${EXCLUDED_PALLETS[@]}"; do
-      if [ "${exluded}" == "${pallet}" ]; then
+      if [ "${exluded}" = "${pallet}" ]; then
           return 0
       fi
   done
-  
+
   return 1
 }
 
 # Delete the error file before each run.
-rm -f "${ERR_FILE}"
+rm -f "${ERR_FILE:?err_unset}"
 
 # Benchmark each pallet.
 for PALLET in "${PALLETS[@]}"; do
@@ -130,12 +141,12 @@ for PALLET in "${PALLETS[@]}"; do
     continue
   fi
 
-  PALLET_NAME="$(echo "${PALLET}" | tr '_' '-')"
+  PALLET_NAME="$(tr '_' '-' <<< "${PALLET}")"
   MODULE_NAME="${PALLET}.rs";
   WEIGHT_FILE="${WEIGHTS_FOLDER}/${MODULE_NAME}"
-  echo "[+] Benchmarking $PALLET with weight file $WEIGHT_FILE";
+  echo "[+] Benchmarking $PALLET with weight file $WEIGHT_FILE"
 
-  OUTPUT=$( \
+  OUTPUT="$( \
     SOURCE_ROOT="${SOURCE_ROOT}" \
     WEIGTH_OUT_PATH="${WEIGHT_FILE}" \
     SKIP_BUILD="true" \
@@ -145,47 +156,44 @@ for PALLET in "${PALLETS[@]}"; do
     BM_STEPS="${BM_STEPS}" \
     BM_REPEAT="${BM_REPEAT}" \
     BM_HEAP_PAGES="${BM_HEAP_PAGES}" \
+    BASE_PATH_ARG="${BASE_PATH_ARG}" \
     "${BENCH_SH}" "${PALLET_NAME}" 2>&1
-  )
-
-  if [ $? -ne 0 ]; then
-    echo "$OUTPUT" >> "$ERR_FILE"
-    echo "[-] Failed to benchmark $PALLET. Error written to $ERR_FILE; continuing..."
-  fi
+  )" || {
+    echo "$OUTPUT" >> "$ERR_FILE";
+    echo "[-] Failed to benchmark $PALLET. Error written to $ERR_FILE; continuing...";
+  }
 done
 
-if [[ "${ENABLE_OVERHEAD}" == "true" ]];
-then 
+if [ "${ENABLE_OVERHEAD:-}" = "true" ]; then
   # Update the block and extrinsic overhead weights.
   echo "[+] Benchmarking block and extrinsic overheads..."
-  OUTPUT=$(
+  # shellcheck disable=SC2086
+  OUTPUT="$(
     ${NH_NODE} benchmark overhead \
     --chain=dev \
     --weight-path="${WEIGHTS_FOLDER}" \
     --header="${CODE_HEADER}" \
     --warmup=10 \
-    --repeat=100 2>&1
-  )
-
-  if [ $? -ne 0 ]; then
-    echo "$OUTPUT" >> "$ERR_FILE"
-    echo "[-] Failed to benchmark the block and extrinsic overheads. Error written to $ERR_FILE; continuing..."
-  fi
-else 
+    --repeat=100 \
+    ${BASE_PATH_ARG} 2>&1
+  )" || {
+    echo "$OUTPUT" >> "$ERR_FILE";
+    echo "[-] Failed to benchmark the block and extrinsic overheads. Error written to $ERR_FILE; continuing...";
+  }
+else
   echo "[+] Skipping - Benchmarking block and extrinsic overheads..."
 fi
 
-if [[ "${ENABLE_MACHINE}" == "true" ]];
-then 
+if [ "${ENABLE_MACHINE:-}" = "true" ]; then
   echo "[+] Benchmarking the machine..."
-  OUTPUT=$(
-    ${NH_NODE} benchmark machine --chain=dev 2>&1
-  )
-  if [ $? -ne 0 ]; then
+  # shellcheck disable=SC2086
+  OUTPUT="$(
+    ${NH_NODE} benchmark machine --chain=dev ${BASE_PATH_ARG} 2>&1
+  )" || {
     # Do not write the error to the error file since it is not a benchmarking error.
-    echo "[-] Failed the machine benchmark:
-  $OUTPUT"
-  fi
+    echo -e "[-] Failed the machine benchmark:\n";
+  }
+  echo "${OUTPUT}"
 else
   echo "[+] Skipping - Benchmarking the machine..."
 fi
